backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
httpx
PyJWT
